requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
pymongo==4.7.3
aiohttp==3.9.5